requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
motor>=3.3.0
pydantic>=2.4.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
beanie>=1.22.6